Refactor: split page.tsx into section and reusable components
Extract monolithic page into:
- data/site-data.ts: articles, companies, techStack, nav/footer links
- components/ui/: ArrowIcon, BrandBackground, SectionHeader (reusable)
- components/sections/: Navbar, Hero, Services, Expertise, TechStackBar,
  Clients, Articles, ContactCTA, Footer
- ServiceCard extracted as reusable within Services

Simplify tailwind content path to scan all src/ files.
