+++ src/components/ui/BrandBackground.tsx
@@ -1,0 +1,66 @@
+export default function BrandBackground({
+	showMark = false,
+}: {
+	showMark?: boolean;
+}) {
+	return (
+		<div className="absolute inset-0 overflow-hidden pointer-events-none">
+			{/* Soft cyan glow */}
+			<div className="absolute top-1/4 right-1/3 w-[600px] h-[600px] bg-secondary/[0.04] rounded-full blur-[120px]" />
+
+			<svg
+				className="absolute inset-0 w-full h-full"
+				xmlns="http://www.w3.org/2000/svg"
+				preserveAspectRatio="none"
+			>
+				{/* Horizontal parallel lines (left side) */}
+				{Array.from({ length: 10 }, (_, i) => (
+					<line
+						key={`hl-${i}`}
+						x1="0"
+						y1={`${28 + i * 5}%`}
+						x2="50%"
+						y2={`${28 + i * 5}%`}
+						stroke="#00a8b7"
+						strokeWidth="1"
+						opacity={0.06 + (i % 3 === 0 ? 0.03 : 0)}
+					/>
+				))}
+
+				{/* Concentric arcs from right */}
+				{Array.from({ length: 7 }, (_, i) => (
+					<circle
+						key={`arc-${i}`}
+						cx="92%"
+						cy="65%"
+						r={`${10 + i * 7}%`}
+						fill="none"
+						stroke="#00a8b7"
+						strokeWidth="1"
+						opacity={0.1 - i * 0.01}
+					/>
+				))}
+			</svg>
+
+			{/* Large R logo mark watermark */}
+			{showMark && (
+				<svg
+					className="absolute -right-10 top-1/2 -translate-y-1/2 w-[500px] h-[420px] opacity-[0.03]"
+					viewBox="0 0 109.8806 84.642"
+					xmlns="http://www.w3.org/2000/svg"
+				>
+					<path
+						d="M9.737,54.8609c0,5.3831,2.1004,10.4463,5.9395,14.2437,3.7835,3.7418,8.7772,5.8005,14.1047,5.8005h.1948l13.4092-.0417c3.7556,1.5857,7.72,3.7418,11.7678,6.4125,1.7665,1.1545,3.547,2.2535,5.3692,3.2689l-30.4906.0974h-.2782c-7.873,0-15.3148-3.0463-20.9344-8.6242C3.1297,70.3843,0,62.8591,0,54.8609c0-16.4277,13.3536-29.7952,29.7812-29.7952h41.7716c-.8346,1.4327-1.3075,3.1019-1.3075,4.8685s.4729,3.4357,1.3075,4.8684H29.7812c-11.0584,0-20.0442,8.9997-20.0442,20.0582Z"
+						fill="white"
+					/>
+					<path
+						d="M109.2908,84.5446h-4.271c-11.3227,0-22.4503-2.4899-32.7854-7.2749-1.6553-.765-3.2828-1.5857-4.8963-2.476-1.9336-1.0432-3.8253-2.1838-5.6892-3.4079-5.9396-3.8948-14.0769-8.3599-22.6594-10.4047,1.3076-1.6413,2.0726-3.7278,2.0726-5.9813,0-1.7804-.4729-3.4496-1.3214-4.8963h40.2276c5.5359,0,10.6826-2.1978,14.5217-6.1621,3.8391-3.9782,5.8423-9.2084,5.6477-14.7445-.3892-10.7246-9.6951-19.4322-20.7536-19.4322l-79.3841-.0278V0l79.3841.0278c16.2471,0,29.9204,12.9223,30.4769,28.8214.2919,8.193-2.6707,15.9547-8.3602,21.8525-5.6887,5.8839-13.3396,9.1388-21.5325,9.1388h-18.5003c1.8361,1.0571,3.6722,2.1977,5.5361,3.4218,11.4897,7.5531,24.6206,11.5452,37.9601,11.5452h4.8684l-.542,9.737Z"
+						fill="white"
+					/>
+					<circle cx="31.3252" cy="54.9999" r="9.737" fill="white" />
+					<circle cx="79.9822" cy="29.9342" r="9.737" fill="white" />
+				</svg>
+			)}
+		</div>
+	);
+}
