+++ src/data/site-data.ts
@@ -1,0 +1,59 @@
+export const articles = [
+	{
+		title: 'My First Open Source Contribution',
+		url: 'https://medium.com/@francois-roget/my-first-open-source-contribution-5707aebc0408',
+		image:
+			'https://miro.medium.com/v2/resize:fit:4800/format:webp/1*x3l55V89muzuP3oOOSxKuw.jpeg',
+		tags: ['Open Source', 'Github'],
+	},
+	{
+		title: 'How to Conditionally Render React UI Based on User Permissions',
+		url: 'https://medium.com/@francois-roget/how-to-conditionally-render-react-ui-based-on-user-permissions-7b9a1c73ffe2',
+		image:
+			'https://miro.medium.com/v2/resize:fit:4800/format:webp/1*uJ7uijA2eurqLs_XxXjO-w.jpeg',
+		tags: ['React', 'Permissions'],
+	},
+	{
+		title: 'Accelerate Your Prototypes Development by Using a Monorepo',
+		url: 'https://medium.com/stackademic/accelerate-your-prototypes-development-by-using-a-monorepo-2ecac78e9087',
+		image:
+			'https://miro.medium.com/v2/resize:fit:1400/format:webp/1*eJhDX2imHUMrTXYcEeR_Qg.jpeg',
+		tags: ['Monorepo', 'Prototyping'],
+	},
+	{
+		title: 'How to Build a Dynamic Micro-Frontend Architecture in React',
+		url: 'https://medium.com/@francois-roget/how-to-build-a-dynamic-micro-frontend-architecture-in-react-95ce548cb775',
+		image:
+			'https://miro.medium.com/v2/resize:fit:1400/format:webp/1*BseOQyheuz2bqJtLss60tg.png',
+		tags: ['React', 'Micro Frontends'],
+	},
+];
+
+export const companies = [
+	{ name: 'Eurocontrol', logo: '/logos/Eurocontrol.png' },
+	{ name: 'Be.Brussels', logo: '/logos/Be_Brussels.png' },
+	{ name: 'APB', logo: '/logos/apb.png' },
+	{ name: 'Tunz', logo: '/logos/Tunz.webp' },
+	{ name: 'Ingenico', logo: '/logos/Ingenico.png' },
+	{ name: 'Worldline', logo: '/logos/Worldline.png' },
+	{ name: 'Paynovate', logo: '/logos/Paynovate.png' },
+	{ name: 'Keytrade Bank', logo: '/logos/Keytrade.png' },
+];
+
+export const techStack = [
+	{ name: 'React', color: 'bg-[#61DAFB]' },
+	{ name: 'TypeScript', color: 'bg-[#3178C6]' },
+	{ name: 'Java', color: 'bg-[#ED8B00]' },
+	{ name: 'PHP', color: 'bg-[#777BB4]' },
+	{ name: 'Nx', color: 'bg-[#143055]' },
+];
+
+export const navLinks = ['Services', 'Expertise', 'Clients', 'Articles'];
+
+export const footerLinks = [
+	'Services',
+	'Expertise',
+	'Clients',
+	'Articles',
+	'Contact',
+];
